rag-backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
torchaudio
fastapi
uvicorn
python-multipart
pymupdf
faiss-cpu
sentence-transformers
huggingface_hub
sentencepiece
opencv-python-headless
boto3
markdown
requests
hf_xet
av
openai
pydantic
python-dotenv
pandas
openpyxl
python-docx
pdf2image
pytesseract
numpy
spaces
flashrank
diskcache
reportlab
markdown2
